Do no initialize J9ModronHeapFreeList with memset
This is a non-trivial class, and should not be initialized with memset.
A constructor has been created. Additionally, the actual class was
renamed from `J9ModronHeapFreeList` to `J9ModronFreeList`, and the
typedef was removed.  It was likely a spelling mistake to name the
struct and typedef different things.  J9ModronFreeList was chosen
because it was more common in the rest of the code.

Signed-off-by: Andrew Young <youngar17@gmail.com>