Remove the unused dlfcn.h header from jitbuilder
The header `dlfcn.h` - dynamic linking - the file defines a set of
macros and functions that allows runtime dynamic library loading.
The header is included in the sources of `jitbuilder/release` even
though no defined macros and functions are used there. The commit
removes the includes from the `jitbuilder/release` C++ files.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>