Clang: disable unused private field and function
There are many warnings `unused-private-field` about the
`OMRNode::NodeExtensionStore::_flag` appear during compilation
with clang as well as some false positive warnings `unused-function`
about the `isJ9()` function (the function is used!) do. To suppress the
warnings, the `-no-unused-private-field -no-unused-function` command
line was added to clang in the MSVC toolchain configuration.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>