Fix container detection with systemd
Up until now, `isRunningInContainer` assumed the process was
operating in a container if init (pid 1) wasn't in the root
cgroup in any subsystem. This is broken in systemd, as it
encapsulates pid 1 in a special "init.scope" group.

Added the exception of cgroup "init.scope" for systemd
version 226 and onwards.

Signed-off-by: Mohammad Ali Nikseresht <anikser@ibm.com>