Fix tril's Xcall related tests
Tril parses hexadecimal numbers if and only if they start with the `0x`
prefix. Unfortunately, some C/C++ compilers put the prefix for the `%p`
format specifier (so, they translate `%p` to something like
`0x0123456789ABCDEF`) while others don't. If the `0x` prefix is absent
before the address attribute value of the `Xcall` node, tril
interprets leading digits as a decimal number and the rest starting from
not a digital symbol (A,B,C,D,E or F) as an anonymous attribute. As a
result, the formed `Xcall` node will try to invoke a function by a wrong
address and an Access Violation Error appears.

The commit fixes the problem explicitly adding the `0x` prefix before the
format specifier in all code for `Xcall` tests. Output for the ilgen trace
of the `Xcall` opcode is fixed too.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>