Use compiler defined size_t instead of std::size_t
For extensible classes such as IlGeneratorMethodDetails when overridden
in downstream projects we may encounter compilation issues as the
extended class will look for a suitable delete operator which should be
compatible with the types on the new operator.

size_t and std::size_t however are not compatible for all compilers
(gcc) and will cause errors complaining that no suitable delete
operator was found for the extensible class.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>