Add cgroup limit checking in omrsysinfo_get_number_CPUs_by_type
Changes `OMRPORT_CPU_BOUND` to check if the CPU subsystem is enabled,
indicating that cgroup limits be used. If so, then `OMRPORT_CPU_BOUND`
return value is capped at an "equivalent" number of processors equal
to `(cpu.cfs_quota_us / cpu.cfs_period_us)` rounded to the nearest
natural number (this "equivalent number of processors" calculation
may be subject to change in the future).

This change only affects Linux and addresses problems discussed in
eclipse/openj9#1166.

This issue depends on #2605.

Issue: eclipse/openj9#1166
Signed-off-by: Mohammad Ali Nikseresht <anikser@ibm.com>