Disable HPR GRA unless register pressure simulation is enabled
Register pressure simulation is a prerequisite for HPR GRA because GRA
and local RA need to make consistent choices and register pressure
simulation is the only part of GRA that is HPR aware. As concrete
examples, among others, consider the following:

1. A collected reference coming in as a parameter

In this case GRA needs to know that on 64-bit such a register candidate
should not be considered for HPRs, since they are really 32-bit
registers. However GRA does not know anything about this. It is the
register pressure simulation algorithm [1] that coordinates with the
codegen on whether collected references are HPR elligible.

2. A valid HPR candidate is being used as a return value

In this case GRA needs to be aware of the choices local RA will make.
Because a value feeds into a return point of a method local RA must
enforce that the virtual register corresponding to the return value is
64-bit [3]. Otherwise the high order half of the register may get
locally allocated to an HPR spill, and of course this would not be
valid. As such GRA must know this fact and it must not globally
allocate the return value to an HPR, otherwise we will get into an
impossible scenario where local RA is forced to coerce a 64-bit GPR
(the return value) into a 32-bit HPR.

The register pressure algorithm is once again aware of these
interactions and prevents such values feeding into return points from
being globally HPR allocated [2].

[1] https://github.com/eclipse/omr/blob/9d1d8cf3048781bc6d87e6a1079167586cc5aa4d/compiler/codegen/CodeGenRA.cpp#L2691-L2702
[2] https://github.com/eclipse/omr/blob/9d1d8cf3048781bc6d87e6a1079167586cc5aa4d/compiler/codegen/CodeGenRA.cpp#L2889-L2903
[3] https://github.com/eclipse/omr/blob/9d1d8cf3048781bc6d87e6a1079167586cc5aa4d/compiler/z/codegen/ControlFlowEvaluator.cpp#L1098-L1102

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>