Add signalMap for win64amd
signalMap includes the list of signals supported by a platform.

OMRPORT_SIG_FLAG_SIGALLASYNC is a bit-mask of all the asynchronous
signals that OMR signal library supports on all platforms.

OMRPORT_SIG_FLAG_SIGALLSYNC is a bit-mask of all the synchronous signals
that OMR signal library supports on all the platforms.

signalMap has been added for win64amd. It specifies the signals which
are supported on win64amd.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>