Add a new flag for testing JitBuilder
This flag is enabled by default if OMR_JITBUILDER is enabled but
it is disbaled on Windows for now since the tests do not compile
due to c++11 features which are not available on Visual Studio 11
2012.

Move ToIlType into extra tests since it requires c++11 features
that are not available on all of our compilers.

Disable running of the JitBuilder samples on 32 bit since they
currently crash on 32 bit.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>