Function may return without exiting the monitor [win32]
In win32, omrsig_startup may return without exiting the globalMonitor if
omrthread_tls_alloc fails.

A static function, initializeSignalTools, is added to initialize signal
tools. It returns 0 on success and non-zero otherwise. It properly free
allocated tools in case of failure.

A call to initializeSignalTools in omrsig_startup resolves the issue of
returning without exiting the globalMonitor. 

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>