Fix counting of GPRs and HPRs on Z codegen
Currently HPRs are not propperly tagged with TR_HPR register kind and
are instead labeled as TR_GPR. This produces incorrect counts for the
number of GPR vs. HPR dependencies that we have. We use a workaround
here to circumvent the issue but a larger change is needed to properly
tag the register kinds and eliminate this ugly code.

We also handle removed GPR / HPR dependencies in
checkRegisterDependencyDuplicates in which we don't actually remove the
HPR dependency but simply set the target real register to  be
TR::RealRegister::NoReg. This is also a hack that needs to be cleaned
up, i.e. we shouldn't  be creating the duplicate register dependency in
the first place.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>