Add OMRMethodBuilder fields to support MethodBuilder inlining
To support MethodBuilder inlining, a MethodBuilder object needs to
know its inline depth, the return builder to which all Return
operations should Goto, and the name of the symbol into which any
return value has been (consistently for all paths) stored.

With the addition of a return builder, the MethodBuilder can also
discover the caller MethodBuilder by asking the return builder for
its owning MethodBuilder.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>