Add matched user-defined placement delete operator
When memory is allocated for an object with operator new, the object's
constructor is called. If the constructor throws an exception, any
memory that was allocated for the object should be deallocated. This
cannot take place unless an operator `delete` function exists that matches
the operator `new`.

The compiler code generates a lot of **C4291** warnings because no placement
form of operator `delete` has been defined that matches the placement form
of operator `new`.

The detailes are described on MSDN:
https://msdn.microsoft.com/en-us/library/cxdxz3x6.aspx

The commit adds missed placement forms of operator `delete` to the
compiler code.

For the optimizator files bellow:

 * `OMROptimization.hpp`
 * `DataFlowAnalysis.hpp`
 * `UseDefInfo.hpp`
 * `SinkStores.hpp`
 * `ValueNumberInfo.hpp`
 * `OptimizationPolicy.hpp`
 * `OptimizationUtil.hpp`
 * `OMROptimizationManager.hpp`

the placement `delete` operator

`static void  operator delete(void *ptr, TR::Allocator a)`

does nothing: @dsouzai has confirmed these objects exist only during
a compilation. If there is an exception thrown during construction,
the compilation will be aborted, and all memory associated with that
compilation will get freed (when `TR::SegmentAllocator &scratchSegmentProvider`
goes out of scope).

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>