Fix ranges used by findAvailableMemoryBlockNoMalloc
The caller provides start and end addresses that constrain a non-null
address that can be returned by the function. Internally, however,
the function operates on ranges which are inclusive of the space
requested by the caller: the allowed range must be initialized taking
this difference into account by adding the requested block size to the
end address.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>