Add feature to ignore spin thread bound
In OMR, a user can set lib->maxSpinThreads to 0. In the current
implementation, no threads will spin if lib->maxSpinThreads is set to 0.
This is a redundant feature since spinning can also be disabled by
setting lib->defaultMonitorSpinCount1/2/3 to 1.

So, the behavior for lib->maxSpinThread = 0 is re-purposed. If
lib->maxSpinThreads is set to 0, then an upper bound won't be enforced
on spin threads, and unlimited spin threads will be allowed.

This is an experimental feature. A consumer of OMR can enable this
feature by setting lib->maxSpinThread to 0 via
omrthread_global("maxSpinThreads"). It will be used to create heuristics
for locking workloads which benefit from no bound on the number of spin
threads.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>