Fix JitBuilder Call to return the right type for Int8 and Int16
Currently if a function returns a type smaller than Int32 the
returned value will be of type Int32 and not Int16 or Int8. I
added some tests that will assert in JitBuilder if the type of
the return value does not match the declaration. This is achieved
by adding the result to a Const<type> of the appropriate type.

The solution is to convert the result if callNode->getDataType()
does not match the defined return type. To make the change simpler
I modified the parameters of convertTo to take a TR::DataType
instead of a TR::IlType.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>