Introduce CodeCache::initialize() without CodeCacheHashEntrySlab parm
As part of the work to stage the changes for #2832, this PR introduces a
new API for the OMR::CodeCache::initialize() function that does not
accept a CodeCacheHashEntrySlab parameter.  Instead, it will allocate the
slab as part of the initialization process thereby encapsulating the
trampoline requirements within the CodeCache.

This single change is being introduced to allow downstream projects to
adapt to the changes proposed in #2832.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>