Update PPA format to RRF2
According to PoPs the PPA instruction is an RRF-c format instruction
with no third register and a mask, hence it should be labeled as an
RRF2 instruction.

Note that the .format field in the table modified is only used as a
reference and has no bearing on any execution path.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>