Improve VirtualMachineRegister* constructors
Currently in OMRVirtualMachineRegister and
OMRVirtualMachineRegisterInStruct require the value being
simulated by the register to be a PointerTo(PointerTo(primitive)).
A virtual machine may want to simulate a PC or an SP and in most
common cases these are a primitive type and a pointer type
respectively. The OMRVirtualMachineRegister* classes have been
updated to account for these. To be able to properly assert
conditions OMRVirtualMachineRegister.cpp was created for methods
that contain TR_ASSERT calls.

I corrected OperandStackTest to properly define it's struct and
also cleaned up OperandArrayTest.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>