Add stubs for omrsig_is_signal_ignored
omrsig_is_signal_ignored checks if the signal is ignored by the OS. If
the signal is ignored, isSignalIgnored is set to TRUE. Otherwise,
isSignalIgnored is set to FALSE.

If an error is encountered during evaluation, OMRPORT_SIG_ERROR is
returned. 0 is returned on success. If OMRPORT_SIG_ERROR is returned,
then the value of isSignalIgnored should be ignored.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>