Add new API for RRF Instructions
RRF instructions such as PPA15 ignore the register fields
in their encoding. In order to simplify the generation of
these RRF instructions, this commit introduces a new API that
allows the creation of new RRF-C instructions without
the need to provide registers. The code generator will
encode zeros in the instruction's register fields automatically.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>