Fix Struct/Union fieldName resolution
If a struct of union contains fields whose name are a sub string
of another field TypeDictionary will return the wrong field for
one of the names. The fix is to use strcmp instead of strncmp to
ensure the strings are exactly the same.

I have added a simple test for each of these that verifies the
types of each field returned by GetFieldType and GetUnionFieldType
matches the declared type for that field.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>