Update relocation for ClassAddress
Under the option to use the Symbol Validation Manager, the correct way
to acquire the address for classes is to go through the
manager, since it maintains consistency with all the validations as well
as relocations. This commit transmutes the TR_ClassAddress into a
TR_SymbolFromManager relocation record in order to maintain
consistency with the rest of the validations.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>