Make OMR::VirtualMachineState client aware
Client awareness follows the approach used for IlBuilder.

In addition, a special convention is introduced for dealing with virtual
functions (callbacks) that may be overriden on both the client-side and
implementation-side. The functions requiring the use of this convention
are:

* VirtualMachineState::Commit()
* VirtualMachineState::Reload()
* VirtualMachineState::MakeCopy()
* VirtualMachineState::MergeInto()

Functions are overriden on the client side by setting a callback. If the
client chooses not to override the function, the callback can simply
call the corresponding implementation function. Otherwise, the callback
can implement the desired overriding functionality.

An important implication is that calls to the virtual function must
always go through the callback. Therefore, on the implementation side,
an additional function is used that always invokes the callback. Other
parts of the implementation can ensure the correct override is always
invoked by always calling this latter function. Importantly, the
function must not be virtual (nor be overriden). By convention, the
function is given the same name as the function containing the
implementation, but with the first letter being lower-case (the function
containing the implementation starts with a capital letter).

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>