Make OMR::IlBuilder client aware
This commit changes the implementation of the IlBuilder class to enable
support for JitBuilder client APIs and language bindings. The following
key changes are applied:

Add client object pointers to IlBuilder

As part of the enablement of the JitBuilder client API, the notion
that client objects which are 1-1 mapped to JitBuilder objects is
introduced. On the JitBuilder side, this change introduces a field
to all JitBuilder objects to hold a void * pointer to this client
object, which can be set via setClient(). The client object may
ultimately be a value that only has meaning in some language runtime.

Add IlBuilder buildIL client callback support

Defining the client API for JitBuilder includes providing a mechanism for
all the call backs used by JitBuilder to be used by clients via the
client API. This mechanism is a C callback for all virtual functions
that can be overidden. For IlBuilder, that is buildIL() which returns
a bool. The callback can be set via setClientCallback_buildIL() and
will be called with the client object as parameter.

Add client allocator function pointer to IlBuilder

In some cases, JitBuilder objects must be internally constructed. The
_clientAllocator function pointer provides a way for JitBuilder to also
allocate a corresponding client-side object. It is a static field and
must be set by the client API via a call to the setClientAllocator()
static function. Static lifetime is used to void having to construct
JitBuilder objects before setting the field. Also, in most cases, only
a single instance of the field will be needed. However, this also
prevents the use of multiple client APIs in a single process. This may
change once client has had the opportunity to mature.

Add implementation getter function pointer IlBuilder

There are a few cases where JitBuilder must retrieve the implementation
object corresponding to a given client-side object. The _getImpl
function pointer is intended to provide access to a client-defined
function for this purpose. It must be set via a call to setGetImpl().
It is a static field in is therefore bound by similar limitations as
_clientAllocator. Although accessing accessing an implementation object
from a client object is not necessarily needed for IlBuilder, it is
included for generality and consistency with other JitBuilder classes.

Don't use varargs in OMR::IlBuilder::ifAnd() and ifOr()

After separating the client API from the implementation, these two
implementation services should no longer use varargs: the client
API takes care of translating the varargs parameters into arrays
in order to translate client objects to implementation objects.

Move OrphanBytecodeBuilder from MethodBuilder to IlBuilder

Only being able to create BytecodeBuilders from inside a MethodBuilder
is needlessly limiting; IlBuilder is the super class, so let's move it
there for maximum flexibility.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>