Add S390OpProp_SetsCC to instructions missing it
Many instructions have "flag" properties indicating they set specific
condition codes for use in the code generator. However not all
instructions that have a set "flag" property have the
S390OpProp_SetsCC. This is a bug. The S390OpProp_SetsCC should be
present on all instructions as outlined in PoPs. In this commit we add
the S390OpProp_SetsCC property to all instructions missing it.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>