Do not skip generation of LA in aiadd evaluator
When evaluating the following tree we currently have a bug where we do
not generate the actual add because the displacement is too large to be
handled by long displacement instructions:

```
aiadd
  aconst 0x784b8500
  iconst 28
```

This is specific to 31-bit. The reason is because we do not generate
the LA instruction to add the 28 to the previous memory reference. This
bug was introduced in commit 564fb25 as part of #2882. In fact the code
that was removed here is actually an optimization on 31-bit that does
not exist for 64-bit but it really should.

This will be addressed in a separate Issue which will be opened as part
of the PR containing this commit.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>