Implement ILBuilder Negate Service
The OMR compiler supports negate opcodes so it makes sense to
provide a Negate service on ILBuilder. This change also includes
a jitbuildertest for negating.

Enable bneg and sneg on PPC by using the inegEvaluator so that the
ILBuilder service can be available for all sizes on x,p and z.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>