Add S390OpProp_SetsCC to all COMPARE AND BRANCH RELATIVE instructions
The reason we do this is because at binary encoding time we may
determine that the distance to branch is too large to encoding within
such instructions. This is too late however since peephole passes may
have assumed these instructions will not modify the CC.

As such we stay conservative in this case and add the SetCC property
just in case these instructions are later expanded into a two
instruction compare and branch sequence at binary encoding time.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>