Update Tril to accommodate the JitBuilder changes
Tril currently depends on JitBuilder as a compiler backend, though it
does not use its IL generation services. Because it needs to access
functionality that is no longer part of the public JitBuilder API
(i.e. IlInjector), Tril must use the internal JitBuilder API. This
commit therefore updates Tril to follow the changes made to JitBuilder.
Specifically, references to the Jit.hpp header are removed and local
implementations of initializeJit(), shutdownJit(), and
compilerMethodBuilder() are added to the following Tril components:

1. Each individual Tril example
2. The sample compiler/simple_compiler
3. The Tril tests
4. The Tril-based compiler tests

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>