Improve cold block marker under AOT
Under AOT, the compiler assumes some classes are unresolved, even if
they actually aren't. This is because some queries will return NULL if
the class can't be remembered in the SCC. Some optimizations, such as
cold block marker, are affected by whether or not a block contains an
unresolved reference. In AOT, many more blocks are incorrectly marked
cold. This commit attempts to remedy that.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>