Deprecate ShrinkWrapping - Part I
This is the first PR which effectively removes all the ShrinkWrapping code from OMR. The code is removed in such a way that it is easy to see all the coupled parts across the optimizer, codegens, and control.

The optimization currently depends on some cleanup in the downstream OpenJ9 project which has dependencies on this optimizaiton. There will be a subsequent "Part II" PR which will remove some of the shell files so as to not cause build breaks.

Issue: #2107

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>