Remove TRMemory dependence on jitConfig
The provided jitConfig is not used for anything any longer in the TRMemory
hierarchy.  Prepare to remove it.

This commit adds new constructors to the TRMemory hierarchy that do not take
a jitConfig argument to facilitate necessary changes to downstream projects.
The constructors that take a jitConfig will be removed once all known
downstream projects have been modified to use the new constructors without
jitConfig.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>