Make OMR an installable library
This is the start of a series of changes intended to make OMR more
easily consumable.  This change is prototyped using the standalone AVL
and omrsig libraries, although it should be trivial to add support for
other libraries. Some of the ways we hope OMR can be consumed are as
follows:

 1. Installable OMR.  By creating proper install targets for OMR using
    standard paths, it should be trivial to package OMR for any linux
    distribution.
2. CMake User Registry.  With a little bit of care, we can support the
   CMake User Registry.  This means that a user can build OMR anywhere
   on their system, and the build directory will be registered in
   `~/.cmake/packages/Omr`.  This will allow users to consume OMR using
   `find_package(Omr)`, without having to install OMR.
 3. As a subproject.  Having OMR as a subdirectory is the most common
    way to consume OMR.  We intend to continue to support and enhance
    this method of consumption.

We also hope to add versioning support to OMR in the future.

Signed-off-by: Andrew Young <youngar17@gmail.com>