Fold several atomic symbols
This commit creates the following symbols:

- atomicAddSymbol
- atomicFetchAndAddSymbol
- atomicSwapSymbol
- atomicCompareAndSwapSymbol

Because the corresponding "32Bit" and "64Bit" symbols are being used in
the downstream OpenJ9 project a subsequent PR will remove them once the
dependency is alleviated.

The following symbols have been deprecated:

- atomicAdd32BitSymbol
- atomicAdd64BitSymbol
- atomicCompareAndSwap32BitSymbol
- atomicCompareAndSwap64BitSymbol

in favor of the width-less versions. There are no known downstream
projects using the above symbols so folding them is safe.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>