Provide classes to record JitBuilder API calls
Recorder is part of the record and replay functionality.
The ultimate goal for such feature is to remove the JIT from
runtimes and provide it as a service. The recorder part of this
feature, as the name implies, will record the IL at the client side
and send it to be compiled in the server. The server then
replays/parses the IL and compiles it, to then send machine code
back to the client so that it can execute the requested IL. The IL
contains services that have recording support (e.g. Add, Sub, Xor,
Store, Loop, etc. More to come). This commit is dedicated to the
Recorder initial functionality. More details as follows.

JitBuilderRecorder defines the basic structure of a JitBuilder IL
(ljbil) representation. The following files provide implementations
that write this representation to a binary buffer, binary file, text
file, respectively:
OMRJitBuilderRecorderBinaryBuffer
OMRJitBuilderRecorderBinaryFile
OMRJitBuilderRecorderTextFile

To enable replay compilation and a host of other useful facilities,
I have implemented a very simple format that can capture the calls
made to JitBuilder services on builder objects in a readily persisted
(stored to disk, saved in buffers, etc) form.

More details:

This commit defines the basic structure of a JitBuilder IL (ljbil)
representation along with representations as a binary buffer,
binary file, and text file.

The fundamental elements of the jbil representation are (currently):
	String (length then characters)
	Number (8, 16, 32, 64, float, double)
	ID (32-bit)
		Statement
		Type
		Value
		Builder
		Location

An ID is just a number used to identify different Strings, Builder
objects, Values, Types, Locations, and Statements. A std::map is
used to associate an ID with pointers corresponding to these
different categories. Rather than embed pointers into the jbil
output, we embed IDs.

IlBuilder, MethodBuilder, and BytecodeBuilder objects will need
to create some jbil output for each service that is called on
these objects. New objects need to be assigned IDs (by calling
StoreID() ) before they can be referenced by their ID.

Each time an IlBuilder object is created, for example, the NewIlBuilder
can be used to assign it an ID, and that ID can be referenced by
subsequent statements to call new services on it.

Every jbil output starts with (see start() function):
	String(RECORDER_SIGNATURE); // 4 (16 bits) "JBIL"
	Number(VERSION_MAJOR);      // 0 (16 bits)
	Number(VERSION_MINOR);      // 0 (16 bits)
	Number(VERSION_PATCH);      // 0 (16 bits)

This header can be encoded in as little as 12 bytes. This version
can be used to understand how the jbil was produced and enables
processing tools to precisely understand the format of the jbil.

Following the header is a sequence of statements.  There are only
two predefined statements. The first predefined statement assigns
an ID to a string:
	0 "Statement String"

All statements begin with the builder object they apply to.
The predefined statement references the special builder ID
0 which cannot be used by any other builder object. This
statement causes the provided string to be assigned the
next available ID (starting at 1). Subsequent statements
are all of the form:

	<BuilderID> <StatementID> <required values>

Depending on the statement, any number of additional
values may be required, but each statement has a
particular set of values it requires.

The statement strings are defined in StatementNames.hpp
as STATEMENT_<statement name> . For example, STATEMENT_LOAD
represents a call to IlBuilder::Load() while STATEMENT_ADD
represents a call to IlBuilder::Add().

The sequence of statements ends when the second predefined
statement is encountered using the special Builder(1)
object followed by the JBIL_COMPLETE string ("Done");

At the end of the representation, the following trailer is
added (see end() function):
	Builder(1);
	String(JBIL_COMPLETE);      // 4 (16 bits) "Done"

This commit doesn't show how IlBuilder, MethodBuilder, or
BytecodeBuilder can use JitBuilderRecorder to record
jbil; that will come in later commits.

Co-authored-by: Sharon Wang <sharon-wang-cpsc@outlook.com>
Co-authored-by: Mark Stoodley <mstoodle@ca.ibm.com>
Co-authored-by: Andrew Young <youngar17@gmail.com>
Co-authored-by: Charlie Gracie <charlie.gracie@gmail.com>

Signed-off-by: Igor Braga <higorb1@gmail.com>