Get ddrgen working on OSX
The following changes will enable ddrgen to create the generated DDR
classes on OSX.

Create dSYM bundle for OMR shared libraries and dont require .dSYM
extension in parser since the bundle is passed in as the .dbg file.

Dwarfdump displays flag attribute values as a string true/false on OSX
instead of 1 or 0. Parse the value as a string and fall back to looking
for 1 or 0 if it doesn't find either "true" or "false" string.

Parse bit offset field for OSX since the data member location field is
missing if the field is a bit field, and instead we look at the
data_bit_offset field.

Include unknown dies in die tree when parsing dwarfdump output, since
there is an issue with parsing based on last indentation with parent
dies that the parser does not recognize. We then remove the unknown
dies from tree afterwards.

Escape sequences don't work in getmacros script for OSX, so replace
'\t' with character class [[:space:]] and '\U' with bash scripting to
upper case characters. These solutions are much more widely supported
and seem to work with all AIX, BSD, GNU and Windows based tools we use.

Replace // with / in filepath for getmacros script since // comments
out the rest of the line in C/C++.

Signed-off-by: mikezhang <mike.h.zhang@ibm.com>