Set aladd internal pointer flag for VP arraycopy
These nodes are internal pointers and flagging them as such prevents
code generators from including their registers in GC maps.

generateArrayletAddressTree() is left unchanged because it's unclear
whether the leaf element address should be considered an internal
pointer, but this function is currently unused anyhow: it is only called
when generateArraylets(), which with hybrid arraylets is always false.

Fixes #3040

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>