Add awareness of xrdbar operations for Power
Updated evaluateNULLCHKWithPossibleResolve for Power to recognize read
barrier opcodes.

Fixed a bug in the Power version of evaluateNULLCHKWithPossibleResolve
where reference counts were not handled properly. If the child has already
been evaluated, it is not necessary to decrement the reference counts on
the reference node.

Added default xrdbarEvaluators for Power.

Updated copyright dates on modified files.

TM is not currently compatible with read barriers. If read barriers are
needed, TM is disabled.

Issue: #3397
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>