Software RangeCheck  Read Barrier
Renaming flag S/W Evacuate RB to S/W RangeCheck RB. 
Evacuate is an action that is required by RB in Concurrent Scavenger.
It's always executed in software. What has alternative implementation in
S/W vs H/W is the trigger mechanism, which is the range check against
Evacuate area boundaries (or a set of areas).

Temporarily, keeping the old variable and accessor, until downstream
objects are updated to use the new ones.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>