Fold away TR_DisableLongDispNodes as false
By default we let the already existing logic for binary encoding encode
the long displacement instructions inserted for spills and fills.
Rather than manually encoding this we delegate to existing code.

This feature was enabled by default and has been for a very long time.
We do not envision using the fallback code.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>